+++ src/main/java/shop/mtcoding/finalproject/service/OrderService.java
@@ -3,11 +3,9 @@ package shop.mtcoding.finalproject.service;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.data.domain.jaxb.SpringDataJaxb.OrderDto;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -28,8 +26,8 @@ import shop.mtcoding.finalproject.domain.store.StoreRepository;
 import shop.mtcoding.finalproject.domain.user.User;
 import shop.mtcoding.finalproject.domain.user.UserRepository;
 import shop.mtcoding.finalproject.dto.order.OrderReqDto.UpdateToCancleOrderReqDto;
+import shop.mtcoding.finalproject.dto.order.OrderRespDto.OrderHistoryListRespDto;
 import shop.mtcoding.finalproject.dto.order.OrderRespDto.ShowOrderListRespDto;
-import shop.mtcoding.finalproject.util.CustomDateUtil;
 
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
@@ -42,6 +40,7 @@ public class OrderService {
     private final StoreRepository storeRepository;
     private final UserRepository userRepository;
 
+    /* 성진 작업 시작 */
     public OrderHistoryListRespDto 주문내역_목록보기(Long userId) {
         // 1. 해당 유저id로 user정보 셀렉 1셀렉
         User userPS = userRepository.findById(userId).orElseThrow(
@@ -55,36 +54,6 @@ public class OrderService {
         // 4. DTO 응답
         OrderHistoryListRespDto orderHistoryListRespDto = new OrderHistoryListRespDto(orderList);
         return orderHistoryListRespDto;
-    }
-
-    @Getter
-    @Setter
-    public static class OrderHistoryListRespDto {
-        private List<OrderDto> orders;
-
-        public OrderHistoryListRespDto(List<Order> orders) {
-            this.orders = orders.stream().map((order) -> new OrderDto(order)).collect(Collectors.toList());
-        }
-
-        @Getter
-        @Setter
-        public class OrderDto {
-            private String name;
-            private String intro;
-            private String thumbnail;
-            private String deliveryState;
-            private String createdAt;
-
-            public OrderDto(Order order) {
-                this.name = order.getStore().getName();
-                this.intro = order.getStore().getIntro();
-                this.thumbnail = order.getStore().getThumbnail();
-                this.deliveryState = order.getDeliveryStateEnum().getState();
-                this.createdAt = CustomDateUtil.toStringFormat(order.getCreatedAt());
-            }
-
-        }
-
     }
 
     // "/api/order/{userId}"
